Factor out querybuilder/util.ts.
